I've simplified the flex array's migration strategy, while maintaining "correct" behavior.
The old approach brokered what happens when there are competing resize
requests, consolidating them into a single operation. The new strategy
does not do this.  Instead:

1) It forces an ordering of resize requests (so we could do multiple
   ops, instead of just one).

2) If threads see a resize request in front of them that already
   implemented what they want, they consider theirselves done.

Note that there are two functions to kick off resizes now,
flexarray_grow() and flexarray_shrink().  If the array is already past
the requested bounds, we assume some other thread got the job done,
and return doing no work.

With flexarray_grow(), we can often do this cheaply, unless the memory
backing the array is not sufficient to hold the new size, in which
case we need to migrate.

With flexarray_shrink(), the current implementation is always an O(n)
operation (where n is the new number of buckets).  That's because we
care about the case where we shrink and then re-grow.  Particularly,
in the following pseudo-code:

  x = array(200)
  x[100] = 99
  shrink(x, 50)
  grow(x,200)
  y = x[100]

We want y to be told that x[100] is uninitialized, instead of being
told that the value is 99.  Migrating ensures that the old cells do
not get re-used until they are properly re-initialized.

There are other ways we could handle this problem, but they'd take
more overhead in the common cases, and I'd expect that incredibly
frequent downward array resizes are not going to be very common.

For instance, we could have downward resize operations work somewhat
like a migration, where they clear out OLD cells, setting them to
uninitialized.  But then we'd have to have reads check against an
epoch to test whether cells are readable, which would require more
centralized contention around an epoch.

And, all writes would similarly need to make sure that anything they
write is done after any such marking, and then would have to always
write epoch info.

That approach is viable, and wouldn't be all too much slower, but for
the moment, the current approach feels like the superior one.