I've added a kind of FIFO I'm calling a CAPQ.
With capq, you dequeue by asking for the 'top' item, and then doing a
'compare-and-pop' in order to dequeue.

I added this for a very specific reason-- I want to be able to have multiple threads working on a single 'job' in parallel from a queue, and then make sure that no thread accidentally pops work that still needs to be done. It's not as good a general-purpose queue as hq is.

Additionally, I've added an early version of a vector (a flexible array that supports both random access and push/pop). It's passing some simple tests, but most of the code is still untested. However, I'm probably going to spend some significant time on documentation before I come back to this.

Note that with the vector, all operations that might resize the vector are explicitly ordered before they're carried out, using the capq in order to do it.